+++ lib/helpers/fetchEvent.ts
