+++ lib/models/admin/events.ts
@@ -1,0 +1,13 @@
+type AdminEvent = {
+    Id: number;
+    Title: string;
+    Genre: string;
+    Image: string;
+    Description: string;
+    Host: string;
+    Event_Date: string;
+    Event_Time: string;
+    Ticket_Price: number;
+    Tickets_Sold: number;
+    Attendees: number;
+  };
