+++ lib/models/hosts/tickets.ts
@@ -1,0 +1,6 @@
+type TicketType = {
+    ticketName: string;
+    ticketPrice: number;
+    ticketQuantity: number;
+    ticketPerks: string;
+  };
